Create Array Conversion Problem
Takes two arrays. Counts 0 values as an empty slot in arrayA. Checks arrayA for enough empty slots to fit arrayB. If it fits, will merge the arrays, and remove the empty slot values, and sort lowest to highest. If arrayA does not have enough empty slots to hold arrayB, will throw an error.